+++ flatlaf-core/src/main/java/com/formdev/flatlaf/FlatLaf.java
@@ -429,4 +429,8 @@ public abstract class FlatLaf
 
 		return strs;
 	}
+
+	public static boolean isShowMnemonics() {
+		return !SystemInfo.IS_MAC;
+	}
 }